+++ test/breaker.test.js
@@ -1,6 +1,5 @@
 describe('Test breakers', function () {
-  var async = require('async');
-  var MockRedisClient = require('../mocks/mockRedisClient.js');
+  var MockRedisClient = require('./mockRedisClient.js');
   var proxyquire = require('proxyquire');
 
   // Reset before each test
@@ -12,7 +11,7 @@ describe('Test breakers', function () {
 
   beforeEach(function () {
     // Get sharded-redis-client with the redis clients stubbed out
-    ShardedRedis = proxyquire('../../sharded-redis-client.js', {
+    ShardedRedis = proxyquire('../sharded-redis-client.js', {
       redis: MockRedisClient
     });
   });
