+++ test/mockRedisClient.js
